backend/requirements.txt
fastapi
uvicorn
motor
odmantic
python-dotenv
passlib[bcrypt]
pydantic
pydantic[email]
python-jose[cryptography]
uuid
inference_sdk
python-multipart
